Adds output directory option
Allows users to specify an output directory for the extracted frames.

This change introduces a new `--media_folder_full_path` argument
that allows users to define where the extracted frames will be stored.
If the specified directory does not exist, the program attempts to
create it. It also fixes an issue where the log folder was being
created inside the current directory instead of the media folder.